move renderStreaks and renderHabits content to a separate helpers file called profile
